# apps/public/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "ITP Voice Platform",
    "description": "You can try all HTTP operation described in this Swagger spec.\n"
  },
  "host": "192.112.255.30:3000",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json",
    "application/xml"
  ],
  "paths": {
...
    "/accounts?apikey={apiKey}": {
      "get": {
        "tags": [
          "Account"
        ],
        "summary": "List all accounts",
        "description": "List all accounts\n",
        "parameters": [
          {
            "in": "path",
            "name": "apiKey",
            "description": "Api_key",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "An array of Accounts",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Account"
              }
            }
          },
          "default": {
            "description": "Unexpected error"
          }
        }
      },
      "post": {
        "tags": [
          "Account"
        ],
        "summary": "Add accounts",
        "description": "Add accounts\n",
        "parameters": [
          {
            "in": "path",
            "name": "apiKey",
            "description": "Api_key",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "ID of account",
            "schema": {
              "$ref": "#/definitions/Account"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Added account",
            "schema": {
              "$ref": "#/definitions/Account"
            }
          },
          "default": {
            "description": "Unexpected error"
          }
        }
      }
    },
    "/accounts/{param}?apikey={apiKey}": {
      "get": {
        "tags": [
          "Account"
        ],
        "summary": "Get account",
        "description": "Get account\n",
        "parameters": [
          {
            "in": "path",
            "name": "apiKey",
            "description": "Api_key",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "param",
            "description": "parameter to found",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Account",
            "schema": {
              "$ref": "#/definitions/Account"
            }
          },
          "default": {
            "description": "Unexpected error"
          }
        }
      },
      "post": {
        "tags": [
          "Account"
        ],
        "summary": "Update account",
        "description": "Update account\n",
        "parameters": [
          {
            "in": "path",
            "name": "apiKey",
            "description": "Api_key",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "param",
            "description": "parameter to found",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Elements to update",
            "schema": {
              "$ref": "#/definitions/Account"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated account",
            "schema": {
              "$ref": "#/definitions/Account"
            }
          },
          "default": {
            "description": "Unexpected error"
          }
        }
      },
      "delete": {
        "tags": [
          "Account"
        ],
        "summary": "Delete account",
        "description": "Delete account\n",
        "parameters": [
          {
            "in": "path",
            "name": "apiKey",
            "description": "Api_key",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "param",
            "description": "parameter to found",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted account",
            "schema": {
              "$ref": "#/definitions/Account"
            }
          },
          "default": {
            "description": "Unexpected error"
          }
        }
      }
    },
...
    "/accounts/{filter}/get?apikey={apiKey}": {
      "get": {
        "tags": [
          "Get"
        ],
        "summary": "List all",
        "description": "List all\n",
        "parameters": [
          {
            "in": "path",
            "name": "filter",
            "description": "AccountID or realm",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "apiKey",
            "description": "Api_key",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "An array of Elements",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Get"
              }
            }
          },
          "default": {
            "description": "Unexpected error"
          }
        }
      },
      "post": {
        "tags": [
          "Get"
        ],
        "summary": "Add Element",
        "description": "Add Element\n",
        "parameters": [
          {
            "in": "path",
            "name": "filter",
            "description": "AccountID or realm",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "apiKey",
            "description": "Api_key",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "ID of Element",
            "schema": {
              "$ref": "#/definitions/Get"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Added element",
            "schema": {
              "$ref": "#/definitions/Get"
            }
          },
          "default": {
            "description": "Unexpected error"
          }
        }
      }
    },
...
    "/accounts/{filter}/get/{Id}?apikey={apiKey}": {
      "get": {
        "tags": [
          "Get"
        ],
        "summary": "Get By id",
        "description": "Get by id\n",
        "parameters": [
          {
            "in": "path",
            "name": "filter",
            "description": "AccountID or realm",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "apiKey",
            "description": "Api_key",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "Id",
            "description": "Id of element",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Get",
            "schema": {
              "$ref": "#/definitions/Get"
            }
          },
          "default": {
            "description": "Unexpected error"
          }
        }
      },
      "post": {
        "tags": [
          "Get"
        ],
        "summary": "Update By id",
        "description": "Update by id\n",
        "parameters": [
          {
            "in": "path",
            "name": "filter",
            "description": "AccountID or realm",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "apiKey",
            "description": "Api_key",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "Id",
            "description": "id of element to update",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Elements to update",
            "schema": {
              "$ref": "#/definitions/Get"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated Element",
            "schema": {
              "$ref": "#/definitions/Get"
            }
          },
          "default": {
            "description": "Unexpected error"
          }
        }
      },
      "delete": {
        "tags": [
          "Get"
        ],
        "summary": "Delete By id",
        "description": "Delete by id\n",
        "parameters": [
          {
            "in": "path",
            "name": "filter",
            "description": "AccountID or realm",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "apiKey",
            "description": "Api_key",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "Id",
            "description": "id of eleme,t to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted element",
            "schema": {
              "$ref": "#/definitions/Get"
            }
          },
          "default": {
            "description": "Unexpected error"
          }
        }
      }
    },
    "/accounts/{filter}/get/getVariable/{Id}?apikey={apiKey}": {
      "get": {
        "tags": [
          "Get"
        ],
        "summary": "Get By id",
        "description": "Get by id\n",
        "parameters": [
          {
            "in": "path",
            "name": "filter",
            "description": "AccountID or realm",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "apiKey",
            "description": "Api_key",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "Id",
            "description": "Id of element",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Get",
            "schema": {
              "$ref": "#/definitions/Get"
            }
          },
          "default": {
            "description": "Unexpected error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Account": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "enabled": {
          "type": "boolean"
        },
        "key": {
          "type": "string"
        },
        "role": {
          "type": "string"
        },
        "website": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "type": {
          "type": "string",
          "enum": [
            "Residential",
            "Business",
            "Business Enterprise"
          ]
        },
        "accountNumber": {
          "type": "string"
        },
        "accountStatus": {
          "type": "string",
          "enum": [
            "Active",
            "Suspended",
            "Terminated"
          ]
        }
      }
    },
...
    "Get": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "url": {
          "type": "string"
        },
        "variable_name": {
          "type": "string"
        },
        "timeout": {
          "type": "string"
        },
        "account": {
          "type": "string"
        }
      }
    },
...
  }
}